chore: align app and desktop build configs
