+++ packages/desktop/index.html
@@ -17,10 +17,6 @@
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/meslo@1.0.0/meslo.css" />
     <meta name="theme-color" content="#011627" />
     <meta name="theme-color" content="#011627" media="(prefers-color-scheme: dark)" />
-    <!-- iOS PWA meta tags -->
-    <meta name="apple-mobile-web-app-capable" content="yes" />
-    <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent" />
-    <meta name="apple-mobile-web-app-title" content="shuvcode" />
     <meta property="og:image" content="/social-share.png" />
     <meta property="twitter:image" content="/social-share.png" />
   </head>
@@ -41,6 +37,6 @@
     </script>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root" class="flex flex-col h-full bg-background-base"></div>
-    <script src="/src/entry.tsx" type="module"></script>
+    <script src="/src/index.tsx" type="module"></script>
   </body>
 </html>
