+++ packages/desktop/package.json
@@ -1,63 +1,37 @@
 {
   "name": "@opencode-ai/desktop",
-  "version": "1.0.190",
-  "description": "",
+  "private": true,
+  "version": "1.0.191",
   "type": "module",
-  "exports": {
-    ".": "./src/index.ts",
-    "./vite": "./vite.js"
-  },
   "scripts": {
     "typecheck": "tsgo -b",
-    "start": "vite",
+    "predev": "bun ./scripts/predev.ts",
     "dev": "vite",
-    "build": "vite build",
-    "serve": "vite preview"
-  },
-  "license": "MIT",
-  "devDependencies": {
-    "@happy-dom/global-registrator": "20.0.11",
-    "@tailwindcss/vite": "catalog:",
-    "@tsconfig/bun": "1.0.9",
-    "@types/bun": "catalog:",
-    "@types/luxon": "catalog:",
-    "@types/node": "catalog:",
-    "@typescript/native-preview": "catalog:",
-    "typescript": "catalog:",
-    "vite": "catalog:",
-    "vite-plugin-icons-spritesheet": "3.0.1",
-    "vite-plugin-pwa": "1.2.0",
-    "vite-plugin-solid": "catalog:"
+    "build": "bun run typecheck && vite build",
+    "preview": "vite preview",
+    "tauri": "tauri"
   },
   "dependencies": {
-    "@kobalte/core": "catalog:",
-    "@opencode-ai/sdk": "workspace:*",
-    "@opencode-ai/ui": "workspace:*",
-    "@opencode-ai/util": "workspace:*",
-    "@shikijs/transformers": "3.9.2",
-    "@solid-primitives/active-element": "2.1.3",
-    "@solid-primitives/audio": "1.4.2",
-    "@solid-primitives/event-bus": "1.1.2",
-    "@solid-primitives/media": "2.3.3",
-    "@solid-primitives/resize-observer": "2.1.3",
-    "@solid-primitives/scroll": "2.1.3",
+    "@opencode-ai/app": "workspace:*",
     "@solid-primitives/storage": "catalog:",
-    "@solid-primitives/websocket": "1.3.1",
-    "@solidjs/meta": "catalog:",
-    "@solidjs/router": "catalog:",
-    "@thisbeyond/solid-dnd": "0.7.5",
-    "diff": "catalog:",
-    "fuzzysort": "catalog:",
-    "ghostty-web": "0.3.0",
-    "luxon": "catalog:",
-    "marked": "16.2.0",
-    "marked-shiki": "1.2.1",
-    "remeda": "catalog:",
-    "shiki": "3.9.2",
-    "solid-js": "catalog:",
-    "solid-list": "catalog:",
-    "tailwindcss": "catalog:",
-    "virtua": "catalog:",
-    "zod": "catalog:"
+    "@tauri-apps/api": "^2",
+    "@tauri-apps/plugin-dialog": "~2",
+    "@tauri-apps/plugin-opener": "^2",
+    "@tauri-apps/plugin-os": "~2",
+    "@tauri-apps/plugin-process": "~2",
+    "@tauri-apps/plugin-shell": "~2",
+    "@tauri-apps/plugin-store": "~2",
+    "@tauri-apps/plugin-updater": "~2",
+    "@tauri-apps/plugin-http": "~2",
+    "@tauri-apps/plugin-window-state": "~2",
+    "solid-js": "catalog:"
+  },
+  "devDependencies": {
+    "@actions/artifact": "4.0.0",
+    "@tauri-apps/cli": "^2",
+    "@types/bun": "catalog:",
+    "@typescript/native-preview": "catalog:",
+    "typescript": "~5.6.2",
+    "vite": "catalog:"
   }
 }
