--- packages/tauri/src/index.ts
@@ -1,1 +1,0 @@
-import "@opencode-ai/desktop"
