+++ script/publish.ts
@@ -25,7 +25,7 @@ if (!Script.preview) {
   let log = ""
   if (tagExists.exitCode === 0) {
     log =
-      await $`git log ${previousTag}..HEAD --oneline --format="%h %s" -- packages/opencode packages/sdk packages/plugin packages/tauri packages/desktop`.text()
+      await $`git log ${previousTag}..HEAD --oneline --format="%h %s" -- packages/opencode packages/sdk packages/plugin packages/desktop packages/app`.text()
   } else {
     console.log(`Tag ${previousTag} not found locally, skipping changelog generation`)
   }
